+++ app/index.ts
@@ -1,0 +1,8 @@
+import baseDir = require('../base_dir');
+
+export { elementSync, browserSync } from './vars';
+export { ElementFinderSync } from './element-finder-sync';
+export { BrowserSync, TargetLocatorSync, OptionsSync, WindowSync } from './browser-sync';
+export { polledExpect } from './polled-expect';
+export { configure } from './config';
+export { getActiveElement, waitForNewWindow, takeScreenshot, resizeViewport, disallowExpect } from './utility';
