+++ test/spec/test_util.ts
@@ -1,17 +1,17 @@
 import * as fs from 'fs';
 import * as path from 'path';
 
-import { browserSync as browser } from '../../app/protractor_sync';
+import { browserSync } from '../../app/index';
 
 export function injectjQuery() {
-  var jQuery = browser.executeScript(function() {
+  var jQuery = browserSync.executeScript(function() {
     return !!(<any>window).jQuery;
   });
 
   if (!jQuery) {
     var jquerySource = fs.readFileSync(path.join(__dirname, '../../../../node_modules/jquery/dist/jquery.js'), 'utf8');
 
-    browser.executeScript((jquerySource: string) => {
+    browserSync.executeScript((jquerySource: string) => {
       eval(jquerySource);
 
       (<any>window).$.noConflict();
